requirements.txt
pandas~=2.3.2
google-genai~=1.39.1